Move requestToSpeak to StageChannel
